Separate expresion safety analysis to a new file
