+++ compiler/passes/denormalize.cpp
@@ -27,23 +27,17 @@
 #include "LoopStmt.h"
 #include "CForLoop.h"
 #include "WhileStmt.h"
+#include "exprAnalysis.h"
 
 //helper datastructures/types
-
-//maybe more memoization than caching
-std::map<Expr*,bool> safeExprCache;
-std::map<FnSymbol*,bool> globalManipFuncCache;
 
 typedef std::map<SymExpr*, std::pair<Expr*,Type*> > ActualUseDefCastMap;
 
 //prototypes
-bool isExprSafeForReorder(Expr * e);
 bool canPrimMoveCreateCommunication(CallExpr* ce);
 inline bool possibleDepInBetween(Expr* e1, Expr* e2);
 inline bool requiresCast(Type* t);
 inline bool isIntegerPromotionPrimitive(PrimitiveTag tag);
-inline bool isNonEssentialPrimitive(CallExpr* ce);
-bool isSafePrimitive(CallExpr* ce);
 
 bool isDenormalizable(Symbol* sym,
     Map<Symbol*,Vec<SymExpr*>*>& defMap,
@@ -82,7 +76,7 @@ void denormalize(FnSymbol *fn) {
   std::set<CallExpr*> deferredFns;
   ActualUseDefCastMap actualUseDefMap;
 
-  bool cachedGlobalManip = globalManipFuncCache.count(fn);
+  bool cachedGlobalManip = isRegisteredGlobalManip(fn);
 
   forv_Vec(Symbol, sym, symSet) {
 
@@ -99,7 +93,7 @@ void denormalize(FnSymbol *fn) {
       //I think there can be a const but stateful global that is manipulated by
       //the function so, play safe
       if(sym && !sym->isImmediate() && isGlobal(sym)){
-        globalManipFuncCache[fn] = true;
+        registerGlobalManip(fn, true);
         cachedGlobalManip = true;
       }
     }
@@ -121,7 +115,7 @@ void denormalize(FnSymbol *fn) {
 
       //denormalize if the def is safe to move and there is no unsafe function
       //between use and def
-      if(isExprSafeForReorder(def)) {
+      if(exprHasNoSideEffects(def)) {
         if(!possibleDepInBetween(defPar, usePar)) {
           denormalize(def, use, castTo);
         }
@@ -130,7 +124,7 @@ void denormalize(FnSymbol *fn) {
   } // end loop for symbol
 
   if(!cachedGlobalManip) {
-    globalManipFuncCache[fn] = false;
+    registerGlobalManip(fn, false);
   }
 
   //handle deferred actuals
@@ -160,7 +154,7 @@ void denormalizeActuals(CallExpr* ce,
           Expr* defPar = def->parentExpr;
           if(CallExpr* ceTmp = toCallExpr(defPar)) {
             if(!isRecord(ceTmp->get(1)->typeInfo())) { //to preserve pass-by-value
-              if(isExprSafeForReorder(def)) {
+              if(exprHasNoSideEffects(def)) {
                 if(!possibleDepInBetween(defPar, usePar)) {
                   // In C actual evaluation order is not standard, therefore any
                   // defPar that us unsafe for reordering cannot be moved to
@@ -298,93 +292,6 @@ bool isDenormalizable(Symbol* sym,
   return false;
 }
 
-bool isExprSafeForReorder(Expr * e) {
-  if(safeExprCache.count(e) > 0) {
-    return safeExprCache[e];
-  }
-  if(CallExpr* ce = toCallExpr(e)) {
-    if(!ce->isPrimitive()) {
-      FnSymbol* fnSym = ce->theFnSymbol();
-
-      const bool cachedGlobalManip = globalManipFuncCache.count(fnSym);
-
-      if(cachedGlobalManip) {
-        if(globalManipFuncCache[fnSym]) {
-          return false;
-        }
-      }
-      for_formals(formal, fnSym) {
-        if(isReferenceType(formal->typeInfo())) {
-          safeExprCache[e] = false;
-          return false;
-        }
-      }
-
-      std::vector<BaseAST*> asts;
-      collect_asts(fnSym->body, asts);
-      if(!cachedGlobalManip){
-        for_vector(BaseAST, ast, asts) {
-          if (SymExpr* se = toSymExpr(ast)) {
-            Symbol* var = se->var;
-
-            if(!var->isImmediate() &&  isGlobal(var)){
-              safeExprCache[e] = false;
-              globalManipFuncCache[fnSym] = true;
-              return false;
-            }
-          }
-        }
-      }
-
-      //this else if lazily marks a def to be immovable if the
-      //function body has a call to another user function
-      //this can be enhanced by going recursive, but it might be a bit overkill
-      for_vector(BaseAST, ast, asts) {
-        if(CallExpr* ce = toCallExpr(ast)) {
-          if(!isNonEssentialPrimitive(ce)) {
-            safeExprCache[e] = false;
-            return false;
-          }
-        }
-      }
-      //it shouldn't be touching any globals at this point
-      globalManipFuncCache[fnSym] = false;
-    }
-    else {
-      //primitive
-      //if(ce->primitive->isEssential){
-      if(! isSafePrimitive(ce)){
-        safeExprCache[e] = false;
-        return false;
-      }
-      else {
-        //here we have a non essential primitive. But consider following issue:
-        //
-        // tmp = (atomic_read() == 5);
-        //
-        // where `==` is non-essential yet one of its children has side effects
-
-        //I had implemented the following recursion to prevent the issue , but
-        //then realized that it doesn't cause any tests to fail(and I had thought
-        //this would have been a widespread case). Maybe I was wrong?. Further
-        //this has performance impact on compilation time
-
-        /*
-        Expr* e = ce->argList.first();
-        while(e) {
-          if(!isExprSafeForReorder(e)) {//is this recursion safe?
-            return false;
-          }
-          e = e->next;
-        }
-        */
-      }
-    }
-  }
-  safeExprCache[e] = true;
-  return true;
-}
-
 void denormalize(Expr* def, SymExpr* use, Type* castTo) {
   Expr* defPar = def->parentExpr;
 
@@ -483,118 +390,10 @@ bool canPrimMoveCreateCommunication(CallExpr* ce) {
   return false;
 }
 
-inline bool isNonEssentialPrimitive(CallExpr* ce) {
-  return ce->isPrimitive() && isSafePrimitive(ce);
-}
-
-/* List of primitives that we shouldn't be hitting at this point in compilation
-
-    case PRIM_DIV:
-    case PRIM_SIZEOF:
-    case PRIM_USED_MODULES_LIST:
-    case PRIM_STRING_COPY:
-    case PRIM_CAST_TO_VOID_STAR:
-    case PRIM_GET_USER_LINE:
-    case PRIM_GET_USER_FILE:
-    case PRIM_IS_SYNC_TYPE:
-    case PRIM_IS_SINGLE_TYPE:
-    case PRIM_IS_TUPLE_TYPE:
-    case PRIM_IS_STAR_TUPLE_TYPE:
-    case PRIM_NUM_FIELDS:
-    case PRIM_FIELD_NUM_TO_NAME:
-    case PRIM_FIELD_NAME_TO_NUM:
-    case PRIM_FIELD_BY_NUM:
-    case PRIM_IS_UNION_TYPE:
-    case PRIM_IS_ATOMIC_TYPE:
-    case PRIM_IS_REF_ITER_TYPE:
-    case PRIM_IS_POD:
-    case PRIM_COERCE:
-    case PRIM_CALL_RESOLVES:
-    case PRIM_METHOD_CALL_RESOLVES:
-    case PRIM_ENUM_MIN_BITS:
-    case PRIM_ENUM_IS_SIGNED:
-    case PRIM_GET_COMPILER_VAR:k
-*/
-bool isSafePrimitive(CallExpr* ce) {
-  PrimitiveOp* prim = ce->primitive;
-  //if (!prim) return false; // or INT_ASSERT(prim);
-  INT_ASSERT(prim);
-  if (prim->isEssential) return false;
-  switch(prim->tag) {
-    case PRIM_MOVE:
-    case PRIM_SIZEOF:
-    case PRIM_STRING_COPY:
-    case PRIM_GET_SERIAL:
-    case PRIM_NOOP:
-    case PRIM_LOOKUP_FILENAME:
-    case PRIM_REF_TO_STRING:
-    case PRIM_BLOCK_LOCAL:
-    case PRIM_UNARY_MINUS:
-    case PRIM_UNARY_PLUS:
-    case PRIM_UNARY_NOT:
-    case PRIM_UNARY_LNOT:
-    case PRIM_ADD:
-    case PRIM_SUBTRACT:
-    case PRIM_MULT:
-    case PRIM_MOD:
-    case PRIM_LSH:
-    case PRIM_RSH:
-    case PRIM_EQUAL:
-    case PRIM_NOTEQUAL:
-    case PRIM_LESSOREQUAL:
-    case PRIM_GREATEROREQUAL:
-    case PRIM_LESS:
-    case PRIM_GREATER:
-    case PRIM_AND:
-    case PRIM_OR:
-    case PRIM_XOR:
-    case PRIM_POW:
-    case PRIM_MIN:
-    case PRIM_MAX:
-    case PRIM_CAST:
-    case PRIM_TESTCID:
-    case PRIM_GETCID:
-    case PRIM_GET_UNION_ID:
-    case PRIM_GET_MEMBER:
-    case PRIM_GET_MEMBER_VALUE:
-    case PRIM_GET_REAL:
-    case PRIM_GET_IMAG:
-    case PRIM_ADDR_OF:
-    case PRIM_DEREF:
-    case PRIM_PTR_EQUAL:
-    case PRIM_PTR_NOTEQUAL:
-    case PRIM_IS_SUBTYPE:
-    case PRIM_DYNAMIC_CAST:
-    case PRIM_ARRAY_GET:
-    case PRIM_ARRAY_GET_VALUE:
-    case PRIM_WIDE_GET_LOCALE:
-    case PRIM_WIDE_GET_NODE:
-    case PRIM_WIDE_GET_ADDR:
-    case PRIM_IS_WIDE_PTR:
-    case PRIM_CAPTURE_FN_FOR_CHPL:
-    case PRIM_CAPTURE_FN_FOR_C:
-    case PRIM_GET_PRIV_CLASS:
-    case PRIM_GET_SVEC_MEMBER:
-    case PRIM_GET_SVEC_MEMBER_VALUE:
-      return true;
-    case PRIM_UNKNOWN:
-      if(strcmp(prim->name, "string_length") == 0 ||
-          strcmp(prim->name, "object2int") == 0 ||
-          strcmp(prim->name, "real2int") == 0) {
-        return true;
-      }
-      //else fall through
-    default:
-      std::cout << "Unknown primitive : " << prim->name << std::endl;
-      INT_ASSERT(false); // should not be getting those
-      // that are !isEssential and not listed above
-  }
-  return false;
-}
 
 inline bool possibleDepInBetween(Expr* e1, Expr* e2){
   for(Expr* e = e1; e != e2 ; e = getNextExpr(e)) {
-    if(! isExprSafeForReorder(e)) {
+    if(! exprHasNoSideEffects(e)) {
       return true;
     }
   }
